# src/backend/services/auth-service/internal/handler/proto_types.go
// Code generated by protoc-gen-go-grpc. DO NOT EDIT.
// This is a placeholder file. Replace with actual proto-generated code.

package handler

import (
	"context"

	"google.golang.org/grpc"
	"google.golang.org/protobuf/types/known/timestamppb"
)

// Placeholder types - replace with generated proto code

type RegisterRequest struct {
	Email     string
	Password  string
	FirstName string
	LastName  string
	Phone     string
	Role      string
}

type LoginRequest struct {
	Email    string
	Password string
}

type AuthResponse struct {
	User         *User
	AccessToken  string
	RefreshToken string
	ExpiresIn    int64
}

type RefreshTokenRequest struct {
	RefreshToken string
}

type LogoutRequest struct {
	UserId string
}

type LogoutResponse struct {
	Success bool
}

type ValidateTokenRequest struct {
	Token string
}

type ValidateTokenResponse struct {
	Valid  bool
	UserId string
	Role   string
	Email  string
}

type GetProfileRequest struct {
	UserId string
}

type UpdateProfileRequest struct {
	UserId    string
	FirstName *string
	LastName  *string
	Phone     *string
}

type GetUserByIdRequest struct {
	UserId string
}

type VerifyUserRequest struct {
	UserId string
	Status string
	Notes  string
}

type ListUsersRequest struct {
	Page               int32
	PageSize           int32
	Role               *string
	VerificationStatus *string
}

type ListUsersResponse struct {
	Users    []*User
	Total    int32
	Page     int32
	PageSize int32
}

type UploadDocumentRequest struct {
	UserId       string
	DocumentType string
	DocumentData []byte
	Filename     string
}

type UploadDocumentResponse struct {
	DocumentId  string
	DocumentUrl string
}

type GetVerificationStatusRequest struct {
	UserId string
}

type VerificationStatusResponse struct {
	Status    string
	Documents []*IdentityDocument
}

type IdentityDocument struct {
	Id           string
	DocumentType string
	DocumentUrl  string
	UploadedAt   *timestamppb.Timestamp
}

type User struct {
	Id                 string
	Email              string
	FirstName          string
	LastName           string
	Phone              string
	Role               string
	IdentityVerified   bool
	VerificationStatus string
	CreatedAt          *timestamppb.Timestamp
	UpdatedAt          *timestamppb.Timestamp
}

// AuthServiceServer is the server API for AuthService service.
type AuthServiceServer interface {
	Register(context.Context, *RegisterRequest) (*AuthResponse, error)
	Login(context.Context, *LoginRequest) (*AuthResponse, error)
	RefreshToken(context.Context, *RefreshTokenRequest) (*AuthResponse, error)
	Logout(context.Context, *LogoutRequest) (*LogoutResponse, error)
	GetProfile(context.Context, *GetProfileRequest) (*User, error)
	UpdateProfile(context.Context, *UpdateProfileRequest) (*User, error)
	UploadIdentityDocument(context.Context, *UploadDocumentRequest) (*UploadDocumentResponse, error)
	GetVerificationStatus(context.Context, *GetVerificationStatusRequest) (*VerificationStatusResponse, error)
	ValidateToken(context.Context, *ValidateTokenRequest) (*ValidateTokenResponse, error)
	GetUserById(context.Context, *GetUserByIdRequest) (*User, error)
	GetUserByIdAdmin(context.Context, *GetUserByIdRequest) (*User, error)
	VerifyUser(context.Context, *VerifyUserRequest) (*User, error)
	ListUsers(context.Context, *ListUsersRequest) (*ListUsersResponse, error)
}

// UnimplementedAuthServiceServer should be embedded to have forward compatible implementations.
type UnimplementedAuthServiceServer struct{}

func (UnimplementedAuthServiceServer) Register(context.Context, *RegisterRequest) (*AuthResponse, error) {
	return nil, nil
}
func (UnimplementedAuthServiceServer) Login(context.Context, *LoginRequest) (*AuthResponse, error) {
	return nil, nil
}
func (UnimplementedAuthServiceServer) RefreshToken(context.Context, *RefreshTokenRequest) (*AuthResponse, error) {
	return nil, nil
}
func (UnimplementedAuthServiceServer) Logout(context.Context, *LogoutRequest) (*LogoutResponse, error) {
	return nil, nil
}
func (UnimplementedAuthServiceServer) GetProfile(context.Context, *GetProfileRequest) (*User, error) {
	return nil, nil
}
func (UnimplementedAuthServiceServer) UpdateProfile(context.Context, *UpdateProfileRequest) (*User, error) {
	return nil, nil
}
func (UnimplementedAuthServiceServer) UploadIdentityDocument(context.Context, *UploadDocumentRequest) (*UploadDocumentResponse, error) {
	return nil, nil
}
func (UnimplementedAuthServiceServer) GetVerificationStatus(context.Context, *GetVerificationStatusRequest) (*VerificationStatusResponse, error) {
	return nil, nil
}
func (UnimplementedAuthServiceServer) ValidateToken(context.Context, *ValidateTokenRequest) (*ValidateTokenResponse, error) {
	return nil, nil
}
func (UnimplementedAuthServiceServer) GetUserById(context.Context, *GetUserByIdRequest) (*User, error) {
	return nil, nil
}
func (UnimplementedAuthServiceServer) GetUserByIdAdmin(context.Context, *GetUserByIdRequest) (*User, error) {
	return nil, nil
}
func (UnimplementedAuthServiceServer) VerifyUser(context.Context, *VerifyUserRequest) (*User, error) {
	return nil, nil
}
func (UnimplementedAuthServiceServer) ListUsers(context.Context, *ListUsersRequest) (*ListUsersResponse, error) {
	return nil, nil
}

// RegisterAuthServiceServer registers the handler with the gRPC server
func RegisterAuthServiceServer(s *grpc.Server, srv AuthServiceServer) {
	// Placeholder - replace with actual registration
	_ = s
	_ = srv
}